requirements.txt
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# For specific TSFM models
chronos-forecasting>=0.1.0
salesforce-moirai>=0.1.0

# Optional: For experiments
optuna>=3.3.0  # Hyperparameter tuning
wandb>=0.15.0  # Experiment tracking
